Move all the ntoskrnl patches to a separate file and clean things up
