+++ SandboxBootkit/SandboxBootkit.vcxproj.filters
@@ -24,6 +24,9 @@
     <ClCompile Include="EfiUtils.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="PatchNtoskrnl.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Efi.hpp">
@@ -33,6 +36,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="ProcessorBind.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="PatchNtoskrnl.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
